+++ CHANGELOG.md
@@ -1,16 +1,33 @@
 # CHANGELOG
 
+## 1.1.0 - 2024-04-03
+
+* Update protos
+* Enable PML (Predictive Machine Learning) detection and smart feedback
+* Enable bulk mode
+* Enable India region
+* Support for scanning large files (over 2GB)
+
+## 1.0.5 - 2023-12-28
+
+* fix linting issues
+
 ## 1.0.4 - 2023-05-18
+
 * set default timeout_in_seconds to 180 seconds
 
 ## 1.0.3 - 2023-05-10
+
 * Change LICENSE
 
 ## 1.0.2 - 2023-05-10
+
 * Change README.md
 
 ## 1.0.1 - 2023-05-04
+
 * Add scan_buffer() function
-  
+
 ## 1.0.0 - 2023-05-01
+
 * Initial release